// schema.d.ts
export interface operations {
// ...
    "search-multi": {
        parameters: {
            query: {
                query: string;
                include_adult?: boolean;
                language?: string;
                page?: number;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description 200 */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        /**
                         * @default 0
                         * @example 1
                         */
                        page: number;
                        results?: {
                            /**
                             * @default true
                             * @example false
                             */
                            adult: boolean;
                            /** @example /aDYSnJAK0BTVeE8osOy22Kz3SXY.jpg */
                            backdrop_path?: string;
                            /**
                             * @default 0
                             * @example 11
                             */
                            id: number;
                            /** @example Star Wars */
                            title?: string;
                            /** @example en */
                            original_language?: string;
                            /** @example Star Wars */
                            original_title?: string;
                            /** @example Princess Leia is captured and held hostage by the evil Imperial forces in their effort to take over the galactic Empire. Venturesome Luke Skywalker and dashing captain Han Solo team together with the loveable robot duo R2-D2 and C-3PO to rescue the beautiful princess and restore peace and justice in the Empire. */
                            overview?: string;
                            /** @example /6FfCtAuVAW8XJjZ7eWeLibRLWTw.jpg */
                            poster_path?: string;
                            /** @example movie */
                            media_type?: string;
                            genre_ids?: number[];
                            /**
                             * @default 0
                             * @example 78.047
                             */
                            popularity: number;
                            /** @example 1977-05-25 */
                            release_date?: string;
                            /**
                             * @default true
                             * @example false
                             */
                            video: boolean;
                            /**
                             * @default 0
                             * @example 8.208
                             */
                            vote_average: number;
                            /**
                             * @default 0
                             * @example 18528
                             */
                            vote_count: number;
                        }[];
                        /**
                         * @default 0
                         * @example 11
                         */
                        total_pages: number;
                        /**
                         * @default 0
                         * @example 201
                         */
                        total_results: number;
                    };
                };
            };
        };
    };
// ...
}